packages/forge-eval/pyproject.toml
[project]
name = "forge-eval"
version = "0.1.0"
description = "Evaluation metrics: FID, CLIP score, aesthetic scoring"
requires-python = ">=3.11,<3.13"
dependencies = [
    "forge-core",
    "clean-fid>=0.1.35",
    "pydantic>=2.10",
    "scipy>=1.14",
    "torch>=2.5,<2.8",
    "pillow>=11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge"]

[tool.uv.sources]
forge-core = { workspace = true }
